Use labels in Hello World
